+++ tests/integration/models/reddit/test_subreddit.py
@@ -867,6 +867,7 @@ class TestSubredditRelationships(IntegrationTest):
                 'TestSubredditRelationships.moderator__user_filter'):
             moderator = self.subreddit.moderator(redditor='pyapitestuser3')
         assert len(moderator) == 1
+        assert 'mod_permissions' in moderator[0].__dict__
 
     def test_muted(self):
         self.reddit.read_only = False